ca: move SignCertificate to the file where it is used
